feat: add CLI interface, make project CLI-first + MCP
Refactor to CLI-first architecture where all business logic lives in
pure async functions (src/core/operations.ts) consumed by both the CLI
and the MCP server.

- Extract all tool logic from server.ts into src/core/operations.ts
- Add src/core/types.ts with shared result types
- Add src/cli.ts with 16 subcommands (zero external deps)
- Slim server.ts from ~2600 to ~940 lines (thin MCP wrapper)
- Smart entrypoint: CLI commands -> cli, no args -> MCP server (backward compat)
