+++ .pi/plans/22eeea76.md
@@ -1,0 +1,51 @@
+{
+  "id": "22eeea76",
+  "title": "CLI-first + MCP refactor for the-i18n-mcp",
+  "status": "completed",
+  "created_at": "2026-04-28T13:37:45.123Z",
+  "steps": [
+    {
+      "id": 1,
+      "text": "Extract all tool business logic from server.ts into src/core/operations.ts as plain async functions",
+      "done": false
+    },
+    {
+      "id": 2,
+      "text": "Create src/core/types.ts with shared result types for operations",
+      "done": false
+    },
+    {
+      "id": 3,
+      "text": "Create src/cli.ts with subcommands mapping to core operations",
+      "done": false
+    },
+    {
+      "id": 4,
+      "text": "Refactor src/server.ts to be a thin wrapper calling core operations",
+      "done": false
+    },
+    {
+      "id": 5,
+      "text": "Update src/index.ts to route between CLI and MCP modes",
+      "done": false
+    },
+    {
+      "id": 6,
+      "text": "Update package.json bin entries and add any CLI deps if needed",
+      "done": false
+    },
+    {
+      "id": 7,
+      "text": "Build and test that both CLI and MCP modes work",
+      "done": false
+    }
+  ]
+}
+
+Refactor the-i18n-mcp to be CLI-first with MCP as an additional transport.
+
+Architecture:
+- `src/core/operations.ts` — all business logic as pure async functions returning plain objects
+- `src/cli.ts` — CLI entry point with subcommands
+- `src/server.ts` — thin MCP wrapper calling core operations
+- `src/index.ts` — smart router (CLI if args present, MCP if stdio piped)
